requirements.txt
streamlit
google-generativeai
python-dotenv
youtube-transcript-api
langchain  #for ollama
gradio   #for frontend